fix: unify gateway service command environment
