+++ shared/runtime-environment.mjs
@@ -39,6 +39,9 @@ const USER_CONFIG_TEMPLATE = [
   '# 通用 ACP：ACP_COMMAND=your-agent，ACP_ARGS=["--acp"]',
   '# 通用 ACP 如需额外环境变量：QWEN_AUDIO_AGENT_ACP_FORWARD_ENV=NAME_A,NAME_B',
   '',
+  '# 可选：前台 MCP 配置文件的绝对路径；MCP 引用的持久变量也写在本文件中',
+  '# QWEN_AUDIO_FRONTEND_MCP_CONFIG=/absolute/path/to/frontend-mcp.json',
+  '',
   '# 可选日志设置：默认 info、单文件 10 MiB、保留 5 份',
   '# QWEN_AUDIO_LOG_LEVEL=info',
   '# QWEN_AUDIO_LOG_MAX_BYTES=10485760',
